+++ src/Facades/NotTodayHoney.php
@@ -1,0 +1,16 @@
+<?php
+
+namespace Vinksyunit\NotTodayHoney\Facades;
+
+use Illuminate\Support\Facades\Facade;
+
+/**
+ * @see \Vinksyunit\NotTodayHoney\NotTodayHoney
+ */
+class NotTodayHoney extends Facade
+{
+    protected static function getFacadeAccessor(): string
+    {
+        return \Vinksyunit\NotTodayHoney\NotTodayHoney::class;
+    }
+}