+++ CHANGELOG.md
@@ -1,3 +1,7 @@
 # Changelog
 
-All notable changes to `:package_name` will be documented in this file.
+All notable changes to `NotTodayHoney` will be documented in this file.
+
+## 0.1.0 - 2025-12-24
+
+- Initial release of NotTodayHoney honeypot package